+++ CHANGELOG.md
@@ -41,6 +41,7 @@ The major version bump is due to upping the required PHP version from `8.4` to `
 	- `min_count`
 * Added `InjectCache` attribute.
 * Added `InjectSimpleCache`attribute.
+* Added `InjectCrypto`attribute.
 
 #### Changes